Move pipes setup to transform module
